chore: rename client and change export
