+++ apps/laboratory/app/core/universal-provider/page.tsx
@@ -17,7 +17,7 @@ import UniversalProvider from '@walletconnect/universal-provider'
 import base58 from 'bs58'
 import { toHex } from 'viem'
 
-import { AppKit, createAppKit } from '@reown/appkit/basic'
+import { AppKit, createAppKit } from '@reown/appkit/core'
 import { bitcoin, solana } from '@reown/appkit/networks'
 
 import { useChakraToast } from '@/src/components/Toast'
