+++ packages/appkit/src/client/appkit.ts
@@ -33,7 +33,7 @@ import { W3mFrameRpcConstants } from '@reown/appkit-wallet/utils'
 import type { AdapterBlueprint } from '../adapters/ChainAdapterBlueprint.js'
 import { W3mFrameProviderSingleton } from '../auth-provider/W3MFrameProviderSingleton.js'
 import { ProviderUtil } from '../store/ProviderUtil.js'
-import { AppKitCore, type AppKitOptionsWithSdk } from './core.js'
+import { AppKitBaseClient, type AppKitOptionsWithSdk } from './appkit-base-client.js'
 
 declare global {
   interface Window {
@@ -48,7 +48,7 @@ export { AccountController }
 let isInitialized = false
 
 // -- Client --------------------------------------------------------------------
-export class AppKit extends AppKitCore {
+export class AppKit extends AppKitBaseClient {
   static override instance?: AppKit
 
   private authProvider?: W3mFrameProvider
@@ -498,6 +498,7 @@ export class AppKit extends AppKitCore {
       // Selectively import views based on feature flags
       const featureImportPromises = []
       if (features) {
+        // eslint-disable-next-line @typescript-eslint/prefer-optional-chain
         const usingEmbeddedWallet = features.email || (features.socials && features.socials.length)
         if (usingEmbeddedWallet) {
           featureImportPromises.push(import('@reown/appkit-scaffold-ui/embedded-wallet'))
